% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addTSidToLipd}
\alias{addTSidToLipd}
\title{Add TSid where missing in a LiPD file}
\usage{
addTSidToLipd(L, prefix = "add")
}
\arguments{
\item{L}{LiPD file}

\item{prefix}{a prefix to prepend to the randomized TSid}
}
\value{
a Lipd file
}
\description{
Add TSid where missing in a LiPD file
}
