% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{as.lipd}
\alias{as.lipd}
\title{Convert a compatible object into a LiPD Structure}
\usage{
as.lipd(x)
}
\arguments{
\item{x}{a lipd-compatible object}
}
\value{
a LiPD object
}
\description{
Convert a compatible object into a LiPD Structure
}
