% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{as.lipdTs}
\alias{as.lipdTs}
\title{Convert a compatible object into a lipd TS Structure}
\usage{
as.lipdTs(x)
}
\arguments{
\item{x}{a lipd ts compatible object}
}
\value{
a LiPD TS object
}
\description{
Convert a compatible object into a lipd TS Structure
}
