% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{as.lipdTsTibble}
\alias{as.lipdTsTibble}
\title{Convert a compatible object into a lipd TS Tibble object}
\usage{
as.lipdTsTibble(x)
}
\arguments{
\item{x}{a lipd ts tibble-compatible object}
}
\value{
a  a lipd TS Tibble object
}
\description{
Convert a compatible object into a lipd TS Tibble object
}
