% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{as.lipdTsTibbleLong}
\alias{as.lipdTsTibbleLong}
\title{Convert a compatible object into a lipd TS Tibble long object}
\usage{
as.lipdTsTibbleLong(x)
}
\arguments{
\item{x}{a lipd ts tibble long-compatible object}
}
\value{
a  a lipd TS Tibble long object
}
\description{
Convert a compatible object into a lipd TS Tibble long object
}
