% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{as.multiLipd}
\alias{as.multiLipd}
\title{Convert a compatible object into a multiLiPD Structure}
\usage{
as.multiLipd(x)
}
\arguments{
\item{x}{a multiLiPD-compatible object}
}
\value{
a multi LiPD object
}
\description{
Convert a compatible object into a multiLiPD Structure
}
