% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_NA.R
\name{cbind.NA}
\alias{cbind.NA}
\title{handle NAs in cbind}
\usage{
\method{cbind}{`NA`}(..., fill = NA)
}
\arguments{
\item{...}{whatever you're binding}

\item{fill}{unused parameter}
}
\value{
bound columns
}
\description{
handle NAs in cbind
}
