% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cconv}
\alias{cconv}
\title{Neotoma chron conversion metadata}
\format{
## `cconv`
Neotoma conversion metadata for chron
}
\source{
<"1Z44xjSxEDlWnThvYLsHFS9aFAN0FnYh2EdroMs9qe_Q">
}
\usage{
cconv
}
\description{
Neotoma chron conversion metadata
}
\keyword{datasets}
