% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{checkKey}
\alias{checkKey}
\title{Check all values of a given key for validity in appropriate spreadsheet}
\usage{
checkKey(lipdTS, key, keyGeneral)
}
\arguments{
\item{lipdTS}{a lipdTS object}

\item{key}{key name}

\item{keyGeneral}{generalized key name}
}
\value{
invalidDF
}
\description{
Check all values of a given key for validity in appropriate spreadsheet
}
