% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{checkZipMd5}
\alias{checkZipMd5}
\title{Compare the MD5 sums for the queryTable zip between locally stored and current remote}
\usage{
checkZipMd5()
}
\value{
out
}
\description{
Compare the MD5 sums for the queryTable zip between locally stored and current remote
}
