% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{ckeckKey}
\alias{ckeckKey}
\title{Check all values of a given key for validity in appropriate spreadsheet}
\usage{
ckeckKey(lipdTS, key, keyGeneral)
}
\arguments{
\item{key}{}
}
\value{
invalidDF
}
\description{
Check all values of a given key for validity in appropriate spreadsheet
}
