% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{collapse_block}
\alias{collapse_block}
\title{Collapses blocks: calibration, physicalSample, hasResolution,
These follow the regex format of "<key1>_<key2>"
match[[1]][[1]] = full key with underscore and index (ex. "physicalSample_tableName")
match[[1]][[2]] = first key (ex. "physicalSample")
match[[1]][[3]] = second key (ex. "tableName")}
\usage{
collapse_block(entry, l, key, pc)
}
\arguments{
\item{entry}{Time series entry}

\item{l}{Metadata (to append to)}

\item{key}{Key from time series entry}

\item{pc}{paleo or chron? #deprecated}
}
\value{
l: Metadata
}
\description{
Collapses blocks: calibration, physicalSample, hasResolution,
These follow the regex format of "<key1>_<key2>"
match[[1]][[1]] = full key with underscore and index (ex. "physicalSample_tableName")
match[[1]][[2]] = first key (ex. "physicalSample")
match[[1]][[3]] = second key (ex. "tableName")
}
\keyword{internal}
