% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{collapse_block_indexed}
\alias{collapse_block_indexed}
\title{Collapses blocks: funding, publication calibration, interpretation}
\usage{
collapse_block_indexed(entry, l, key)
}
\arguments{
\item{entry}{Time series entry}

\item{l}{Metadata (to add new data to)}

\item{key}{Current key from time series entry}
}
\value{
l Metadata
}
\description{
These follow the regex format of "<key1><idx>_<key2>"

match[[1]][[1]] = full key with underscore and index (ex. "interpretation1_variableDetail")

match[[1]][[2]] = first key (ex. "interpretation)

match[[1]][[3]] = index number (ex. the "1" from "interpretation1")

match[[1]][[4]] = second key (ex. "variableDetail")
}
\keyword{internal}
