% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{collapse_column}
\alias{collapse_column}
\title{Collapse time series column. Compile column entries and place new column in table.}
\usage{
collapse_column(table, entry, pc)
}
\arguments{
\item{table}{Metadata}

\item{entry}{Time series entry}

\item{pc}{paleoData or chronData}
}
\value{
table: Metadata
}
\description{
Collapse time series column. Compile column entries and place new column in table.
}
