% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{collapse_table}
\alias{collapse_table}
\title{Collapse time series section; paleo or chron}
\usage{
collapse_table(d, entry, pc)
}
\arguments{
\item{d}{Metadata}

\item{entry}{Time series entry}

\item{pc}{paleoData or chronData}
}
\value{
d: Metadata
}
\description{
Collapse time series section; paleo or chron
}
