% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{convert_dsid_to_path}
\alias{convert_dsid_to_path}
\title{Convert datasetId to url to lipdverse lipd file}
\usage{
convert_dsid_to_path(dsid, vers = NA)
}
\arguments{
\item{vers}{Optionally specify a dataset version as a character string (e.g. "0_1_2"). Default is NA, which will get the most recent version.}
}
\value{
path to most recent lipd file
}
\description{
Convert datasetId to url to lipdverse lipd file
}
