% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_versions.R
\name{convertVersion}
\alias{convertVersion}
\title{Convert LiPD version structure whenever necessary}
\usage{
convertVersion(D)
}
\arguments{
\item{D}{LiPD Library}
}
\value{
D modified LiPD Library
}
\description{
Convert LiPD version structure whenever necessary
}
\keyword{internal}

