% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-lipd.R
\name{create_mock_neotoma_object}
\alias{create_mock_neotoma_object}
\title{Create a mock neotoma2 site object for testing conversions}
\usage{
create_mock_neotoma_object()
}
\value{
A neotoma2 site object
}
\description{
Create a mock neotoma2 site object for testing conversions
}
