% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-lipd.R
\name{create_test_lipd_object}
\alias{create_test_lipd_object}
\title{Create a basic, valid LiPD object for testing}
\usage{
create_test_lipd_object(dsn = "TestDSN", id = paste0("TestID-", dsn))
}
\arguments{
\item{dsn}{The dataSetName for the object}

\item{id}{The datasetId for the object}
}
\value{
A list formatted as a valid LiPD object
}
\description{
Create a basic, valid LiPD object for testing
}
