% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{createChangelog}
\alias{createChangelog}
\title{Create a changelog by comparing two LiPD files}
\usage{
createChangelog(
  Lold,
  Lnew,
  good.base = c("archiveType", "createdBy", "dataSetName", "dataSource", "notes",
    "originalDataUrl"),
  exclude.paleo = c("paleoData_meanValue12k", "paleoData_mostRecentCompilations",
    "paleoData_medianRes12k", "paleoData_tableName", "paleoData_values"),
  exclude.chron = c("chronData_values")
)
}
\arguments{
\item{Lold}{Old lipd object}

\item{Lnew}{New lipd object}

\item{good.base}{A vector of base level metadata to check}

\item{exclude.paleo}{A vector of paleo metadata to exclude}

\item{exclude.chron}{A vector of chron metadata to exclude}
}
\value{
a changelog data frame
}
\description{
Create a changelog by comparing two LiPD files
}
