% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdManipulation.R
\name{createColumn}
\alias{createColumn}
\title{Create a column in a LiPD object}
\usage{
createColumn(
  L,
  paleo.or.chron = "paleo",
  paleo.or.chron.number = 1,
  table.type = "measurement",
  table.number = 1,
  variableName = NA,
  units = NA,
  values = NA,
  additional.metadata = NA
)
}
\arguments{
\item{L}{a LiPD object}

\item{paleo.or.chron}{is this column for a "paleo" or "chron" table?}

\item{paleo.or.chron.number}{what number is this paleo or chron object (default = 1)}

\item{table.type}{"measurement", "ensemble" or "summary" table? (default = "measurement")}

\item{table.number}{what number is this table object (default = 1)}

\item{variableName}{what is the variableName (e.g. "depth")}

\item{units}{what are the units}

\item{values}{a vector of the values for the column, must be the same length as other columns in the table}

\item{additional.metadata}{an optional named list of additional metadata to add to the column}
}
\value{
updated Lipd object
}
\description{
Create a column in a LiPD object
}
