% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{createMarkdownChangelog}
\alias{createMarkdownChangelog}
\title{Create a Markdown representation of a LiPD changelog}
\usage{
createMarkdownChangelog(L)
}
\arguments{
\item{L}{}
}
\description{
Create a Markdown representation of a LiPD changelog
}
