% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{createProjectChangelog}
\alias{createProjectChangelog}
\title{Create html changelog details and summaries for the differences between two projects}
\usage{
createProjectChangelog(
  Dold,
  Dnew,
  proj,
  projVersOld,
  projVersNew,
  webDirectory = "~/GitHub/lipdverse/html",
  notesTib = NA
)
}
\arguments{
\item{webDirectory}{}
}
\value{

}
\description{
Create html changelog details and summaries for the differences between two projects
}
