% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{createSingleMarkdownChangelog}
\alias{createSingleMarkdownChangelog}
\title{Create a markdown file of a changelog}
\usage{
createSingleMarkdownChangelog(scl)
}
\arguments{
\item{scl}{a changelog for a single file}
}
\value{
markdown
}
\description{
Create a markdown file of a changelog
}
