% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_extract.R
\name{createTimeMetadata}
\alias{createTimeMetadata}
\title{Attach standardized time metadata to a time series entry}
\usage{
createTimeMetadata(current, values, variableName, units)
}
\arguments{
\item{current}{The (shared) time series entry being built}

\item{values}{The values of the identified time column}

\item{variableName}{The variableName of the time column}

\item{units}{The units of the time column}
}
\value{
current with time, timeUnits, timeDatum, timeDirection, timeExponent,
  timeMin and timeMax added
}
\description{
Attach standardized time metadata to a time series entry
}
\keyword{internal}
