% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdManipulation.R
\name{duplicateTable}
\alias{duplicateTable}
\title{Duplicate a table in a LiPD file}
\usage{
duplicateTable(
  L,
  paleo.or.chron = "paleo",
  paleo.or.chron.number = 1,
  table.type = "measurement",
  table.number = 1,
  n.duplicates = 1
)
}
\arguments{
\item{L}{a LiPD object}

\item{paleo.or.chron}{is this a "paleo" or "chron" table?}

\item{paleo.or.chron.number}{what number is this paleo or chron object (default = 1)}

\item{table.type}{"measurement", "ensemble" or "summary" table? (default = "measurement")}

\item{table.number}{what number is this table object (default = 1)}

\item{n.duplicates}{how many duplicates to do you want? (default = 1)}
}
\value{
measurementTable
}
\description{
Duplicate a table in a LiPD file
}
