% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csm.R
\name{expand_csm}
\alias{expand_csm}
\title{Expand csm across a whole time series}
\usage{
expand_csm(ts)
}
\arguments{
\item{ts}{A time series list.}
}
\value{
The time series, with csm expanded.
}
\description{
Expand csm across a whole time series
}
\keyword{internal}
