% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_by_key}
\alias{extract_by_key}
\title{Recursively find all values for a specific key in a nested list.}
\usage{
extract_by_key(lst, key)
}
\arguments{
\item{lst}{The list to search through.}

\item{key}{The character name of the key to find.}
}
\value{
A vector containing all the values found for the specified key.
}
\description{
Recursively find all values for a specific key in a nested list.
}
\examples{
example_list <- list(
  a = 1,
  b = list(
    c = 2,
    target_key = "value1"
  ),
  d = list(
    e = list(
      f = 3,
      target_key = "value2"
    )
  ),
  target_key = "value3",
  g = list(h = 4)
)

extract_by_key(example_list, "target_key")
#> [1] "value3" "value1" "value2"
}
\author{
Gemini, Nick McKay
}
