% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy2.R
\name{fixPubAuthorList}
\alias{fixPubAuthorList}
\title{Fix the structure of author lists}
\usage{
fixPubAuthorList(L)
}
\arguments{
\item{L}{a LiPD object}
}
\value{
a LiPD object
}
\description{
Fix common issue in author lists that get messed up during conversion from lipd.ts.tibble to lipd objects
}
