% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd2neotoma.R
\name{fromOrigLipd}
\alias{fromOrigLipd}
\title{convert an original lipd dataset into a new neotoma site}
\usage{
fromOrigLipd(L)
}
\arguments{
\item{L}{}
}
\value{
neotoma site
}
\description{
convert an original lipd dataset into a new neotoma site
}
