% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma2lipd.R
\name{fromOrigNeotoma}
\alias{fromOrigNeotoma}
\title{Convert lipd back to its original neotoma format}
\usage{
fromOrigNeotoma(L)
}
\arguments{
\item{L}{lipd object}
}
\value{
neotoma site
}
\description{
Convert lipd back to its original neotoma format
}
