% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_extract.R
\name{get_calculate_resolution}
\alias{get_calculate_resolution}
\title{Read the calculateResolution flag set by extractTs}
\usage{
get_calculate_resolution()
}
\value{
logical
}
\description{
Read the calculateResolution flag set by extractTs
}
\keyword{internal}
