% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{get_lipd_version}
\alias{get_lipd_version}
\title{Check what version of LiPD this file is using}
\usage{
get_lipd_version(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
list tmp: Version number and meta
}
\description{
A file with no recognised version key is assumed to be current. Nearly all
such files are recently written ones that simply omitted the key, and the
alternative was a prompt that made reading a directory of files interactive.
}
\keyword{internal}
