% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{getChangelog}
\alias{getChangelog}
\title{Get one instance from a LiPD changelog}
\usage{
getChangelog(L, version = "newest")
}
\arguments{
\item{L}{Lipd object}

\item{version}{which version to get changelog for}
}
\value{
Changelog output
}
\description{
Get one instance from a LiPD changelog
}
