% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_file.R
\name{getData}
\alias{getData}
\title{Retrieve and import csv and jsonld files in the current directory.}
\usage{
getData()
}
\value{
data.list List of data for one LiPD file
}
\description{
Retrieve and import csv and jsonld files in the current directory.
}
\keyword{internal}

