% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma2lipd.R
\name{getGeoNeotoma2}
\alias{getGeoNeotoma2}
\title{get geo info from neotoma}
\usage{
getGeoNeotoma2(site)
}
\arguments{
\item{site}{a single 'site' (not multiple 'sites') object generated by the 'neotoma2' package}
}
\value{
LiPD geo section
}
\description{
get geo info from neotoma
}
\keyword{internal}
