% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma2lipd.R
\name{getPaleoDataNeotoma2}
\alias{getPaleoDataNeotoma2}
\title{get paleoData info from neotoma}
\usage{
getPaleoDataNeotoma2(site)
}
\arguments{
\item{site}{a single 'site' (not multiple 'sites') object generated by the 'neotoma2' package}
}
\value{
LiPD paleoData section
}
\description{
get paleoData info from neotoma
}
\keyword{internal}
