% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{getTimestamp}
\alias{getTimestamp}
\title{Get the timestamp frmo current version of a LiPD file from it's changelog}
\usage{
getTimestamp(L)
}
\arguments{
\item{L}{}
}
\value{

}
\description{
Get the timestamp frmo current version of a LiPD file from it's changelog
}
