% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{getVersion}
\alias{getVersion}
\title{Get the current version of a LiPD file from it's changelog}
\usage{
getVersion(L)
}
\arguments{
\item{L}{}
}
\value{

}
\description{
Get the current version of a LiPD file from it's changelog
}
