% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_indexing.R
\name{indexByName}
\alias{indexByName}
\title{Change index-by-number to index-by-variableName}
\usage{
indexByName(d)
}
\arguments{
\item{d}{LiPD file}
}
\value{
d Modified LiPD file
}
\description{
Change index-by-number to index-by-variableName
}
\keyword{internal}

