% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_indexing.R, R/save_lipds_indexing.R
\name{indexSection}
\alias{indexSection}
\title{Change index-by-number for one section}
\usage{
indexSection(d, keys)

indexSection(d, keys)
}
\arguments{
\item{d}{LiPD metadata}

\item{keys}{Section keys}

\item{d}{LiPD Metadata}

\item{keys}{Section keys}
}
\value{
d Modified LiPD metadata

d Modified metadata
}
\description{
Change index-by-number for one section

Index a single section. Paleo or Chron
}
\keyword{internal}

