% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{initializeChangelog}
\alias{initializeChangelog}
\title{Initialize a changelog for a LiPD object}
\usage{
initializeChangelog(
  L,
  notes = "Starting the changelog",
  curator = Sys.info()[["user"]],
  timestamp = lubridate::now(tzone = "UTC"),
  version = "1.0.0"
)
}
\arguments{
\item{timestamp}{}
}
\value{
A LiPD object with a changelog
}
\description{
This function will create a changelog for a file that doesn't already have one.
}
