% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{is.lipdTsTibble}
\alias{is.lipdTsTibble}
\title{Is this a LiPD TS Tibble object?}
\usage{
is.lipdTsTibble(x)
}
\arguments{
\item{x}{any object}
}
\value{
boolean
}
\description{
Is this a LiPD TS Tibble object?
}
