% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{is.lipdTsTibbleLong}
\alias{is.lipdTsTibbleLong}
\title{Is this a LiPD TS Tibble long object?}
\usage{
is.lipdTsTibbleLong(x)
}
\arguments{
\item{x}{any object}
}
\value{
boolean
}
\description{
Is this a LiPD TS Tibble long object?
}
