% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_extract.R
\name{is_primary_flag}
\alias{is_primary_flag}
\title{Is a column's primary flag truthy?}
\usage{
is_primary_flag(x)
}
\arguments{
\item{x}{value of an isPrimary / primaryAgeColumn field}
}
\value{
logical
}
\description{
Is a column's primary flag truthy?
}
\keyword{internal}
