% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{isValidAll}
\alias{isValidAll}
\title{Are all the parameters valid?}
\usage{
isValidAll(TS, report = TRUE, allKeys = names(standardTables))
}
\arguments{
\item{TS}{lipd-ts object}

\item{report}{report out? T/F}

\item{allKeys}{what keys to check}
}
\value{
boolean
}
\description{
Are all the parameters valid?
}
