% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyStandardization.R
\name{isValidKey}
\alias{isValidKey}
\title{make sure all terms under a controlled key are valied}
\usage{
isValidKey(lipdTS, key)
}
\arguments{
\item{key}{}
}
\value{
invalid keys df
}
\description{
make sure all terms under a controlled key are valied
}
