% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{isValidValue}
\alias{isValidValue}
\title{make sure all terms under a controlled key are valid}
\usage{
isValidValue(lipdTS, key = NA)
}
\arguments{
\item{key}{}
}
\value{
invalid keys df
}
\description{
make sure all terms under a controlled key are valid
}
