% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd2neotoma.R
\name{lipd2neotoma}
\alias{lipd2neotoma}
\title{Convert a LiPD object into a Neotoma site}
\usage{
lipd2neotoma(L)
}
\arguments{
\item{L}{a LiPD object}
}
\value{
a `neotoma2` site object
}
\description{
Converts a LiPD object into a `neotoma2` `site` object, enabling round-tripping
between the LiPD and Neotoma ecosystems. Requires the suggested packages `neotoma2`
and `sf`.
}
\examples{
\dontrun{
L <- readLipd()
site <- lipd2neotoma(L)
}
}
