% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\docType{data}
\name{LIPD_VERSIONS}
\alias{LIPD_VERSIONS}
\alias{LIPD_CURRENT_VERSION}
\title{Known LiPD format versions, and the one assumed when a file states none.}
\format{
An object of class \code{numeric} of length 5.

An object of class \code{numeric} of length 1.
}
\usage{
LIPD_VERSIONS

LIPD_CURRENT_VERSION
}
\description{
Known LiPD format versions, and the one assumed when a file states none.
}
\keyword{internal}
