% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{lipdSummary}
\alias{lipdSummary}
\title{Print a summary of the contents of a LiPD file}
\usage{
lipdSummary(L)
}
\arguments{
\item{L}{a lipd object}
}
\description{
Print a summary of the contents of a LiPD file
}
\seealso{
Other summary: 
\code{\link{formatCitation}()},
\code{\link{lipdTSSummary}()},
\code{\link{matchCols}()},
\code{\link{multiLipdSummary}()},
\code{\link{printCitations}()},
\code{\link{printSummaryData}()}
}
\author{
David Edge
}
\concept{summary}
