% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{lipdTSSummary}
\alias{lipdTSSummary}
\title{Print a summary of the contents of a LiPD TS}
\usage{
lipdTSSummary(
  ts.object,
  time.variable = NULL,
  print.length = 10,
  return.table = FALSE,
  add.variable = NULL
)
}
\arguments{
\item{add.variable}{}
}
\description{
Print a summary of the contents of a LiPD TS
}
\seealso{
Other summary: 
\code{\link{formatCitation}()},
\code{\link{lipdSummary}()},
\code{\link{matchCols}()},
\code{\link{multiLipdSummary}()},
\code{\link{printCitations}()},
\code{\link{printSummaryData}()}
}
\author{
David Edge
}
\concept{summary}
