% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_file.R
\name{loadLipdFiles}
\alias{loadLipdFiles}
\title{Import the data from each csv and jsonld file for given LiPDs}
\usage{
loadLipdFiles(tmp, files_noext)
}
\arguments{
\item{tmp}{Char path to the temp folder in memory}

\item{files_noext}{List of lipd files without extention}
}
\value{
out.list List of data for each lipd file
}
\description{
Import the data from each csv and jsonld file for given LiPDs
}
\keyword{internal}

