% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{matchCols}
\alias{matchCols}
\title{Extends data frames to conserve all columns when merging}
\usage{
matchCols(df1, df2)
}
\arguments{
\item{df1}{}

\item{df2}{}
}
\description{
Extends data frames to conserve all columns when merging
}
\seealso{
Other summary: 
\code{\link{formatCitation}()},
\code{\link{lipdSummary}()},
\code{\link{lipdTSSummary}()},
\code{\link{multiLipdSummary}()},
\code{\link{printCitations}()},
\code{\link{printSummaryData}()}
}
\author{
David Edge
}
\concept{summary}
