% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsons_merge.R
\name{merge_csv_model}
\alias{merge_csv_model}
\title{Merge CSV into each model}
\usage{
merge_csv_model(models, crumbs, csvs)
}
\arguments{
\item{models}{Models to merge}

\item{crumbs}{Crumbs}

\item{csvs}{CSV data}
}
\value{
list models: Metadata
}
\description{
Merge CSV into each model
}
\keyword{internal}
