% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_indexing.R
\name{moveTableUp}
\alias{moveTableUp}
\title{Unindex tables by name. Move to index by number.}
\usage{
moveTableUp(table, pc, tableType)
}
\arguments{
\item{table}{Table data}
}
\value{
table Modified table data
}
\description{
Unindex tables by name. Move to index by number.
}
\keyword{internal}

