% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lipds_main.R
\name{multiParseSave}
\alias{multiParseSave}
\title{Parse library of data records}
\usage{
multiParseSave(D)
}
\arguments{
\item{D}{LiPD data}
}
\value{
none
}
\description{
Parse library of data records
}
\keyword{internal}

