% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc}
\alias{nc}
\title{Neotoma conversion metadata}
\format{
## `nc`
Neotoma conversion metadata
}
\source{
<"1Z44xjSxEDlWnThvYLsHFS9aFAN0FnYh2EdroMs9qe_Q">
}
\usage{
nc
}
\description{
Neotoma conversion metadata
}
\keyword{datasets}
