% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{past}
\alias{past}
\title{Query Table for LiPDverse datasets}
\format{
## `past`
A data frame with 2401 rows and 5 columns:
\describe{
  \item{id}{PaST ID}
  \item{url}{URL To past locator}
  \item{name}{Long name}
  \item{definition}{Definition of term}
  \item{related}{URI to related concepts}
}
}
\source{
<https://github.com/DaveEdge1/lipdverseQuery/raw/main/queryZip.zip>
}
\usage{
past
}
\description{
A table of metadata for the PaST Thesaurus
}
\keyword{datasets}
