% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{print.lipd}
\alias{print.lipd}
\title{short print out of single lipd}
\usage{
\method{print}{lipd}(x, ...)
}
\arguments{
\item{x}{a lipd object}

\item{...}{additional arguments passed to generic print function}
}
\description{
short print out of single lipd
}
