% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{print.lipd_ts}
\alias{print.lipd_ts}
\title{short print out of lipd ts}
\usage{
\method{print}{lipd_ts}(x, ...)
}
\arguments{
\item{x}{a lipd_ts object}

\item{...}{additional arguments passed to generic print function}
}
\description{
short print out of lipd ts
}
