% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{print.lipd_ts_tibble}
\alias{print.lipd_ts_tibble}
\title{short print out of lipd ts tibble}
\usage{
\method{print}{lipd_ts_tibble}(x, ...)
}
\arguments{
\item{x}{a lipd_ts_tibble object}

\item{...}{additional arguments passed to generic print function}
}
\description{
short print out of lipd ts tibble
}
