% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{print.multi_lipd}
\alias{print.multi_lipd}
\title{short print out of multi lipd}
\usage{
\method{print}{multi_lipd}(x, ...)
}
\arguments{
\item{x}{a multi_lipd object}

\item{...}{additional arguments passed to generic print function}
}
\description{
short print out of multi lipd
}
