% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{printCitations}
\alias{printCitations}
\title{print formatted citations to stdout}
\usage{
printCitations(pub)
}
\arguments{
\item{pub}{}
}
\description{
print formatted citations to stdout
}
\seealso{
Other summary: 
\code{\link{formatCitation}()},
\code{\link{lipdSummary}()},
\code{\link{lipdTSSummary}()},
\code{\link{matchCols}()},
\code{\link{multiLipdSummary}()},
\code{\link{printSummaryData}()}
}
\author{
David Edge
}
\concept{summary}
