% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{printModel}
\alias{printModel}
\title{Extract and print model info}
\usage{
printModel(chron.model)
}
\arguments{
\item{chron.model}{model object from chron data of lipd object}
}
\description{
Extract and print model info
}
\author{
David Edge
}
