% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{printSummaryData}
\alias{printSummaryData}
\title{Print summary data from Paleo/Chron measurement tables}
\usage{
printSummaryData(dataIn, run.quiet = FALSE)
}
\arguments{
\item{dataIn}{}
}
\description{
Print summary data from Paleo/Chron measurement tables
}
\seealso{
Other summary: 
\code{\link{formatCitation}()},
\code{\link{lipdSummary}()},
\code{\link{lipdTSSummary}()},
\code{\link{matchCols}()},
\code{\link{multiLipdSummary}()},
\code{\link{printCitations}()}
}
\author{
David Edge
}
\concept{summary}
