% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{put_table}
\alias{put_table}
\title{Put the target table}
\usage{
put_table(d, current, pc, table)
}
\arguments{
\item{d}{Metadata}

\item{current}{Current time series entry}

\item{pc}{paleoData or chronData}

\item{table}{Metadata (to be placed)}
}
\value{
d: Metadata
}
\description{
Put the target table
}
