% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryGet.R
\name{queryGet}
\alias{queryGet}
\title{download a multilipd from lipdverse based on query parameters}
\usage{
queryGet(
  variable.name = NULL,
  archive.type = NULL,
  coord = c(-90, 90, -180, 180),
  age.min = NULL,
  age.max = NULL,
  pub.info = NULL,
  country = NULL,
  continent = NULL,
  ocean = FALSE,
  interpretation = NULL
)
}
\arguments{
\item{variable.name}{}

\item{archive.type}{}

\item{age.min}{}

\item{age.max}{}

\item{pub.info}{}

\item{country}{}

\item{ocean}{}

\item{interpretation}{}

\item{fuzzy}{}

\item{longitude}{}
}
\description{
download a multilipd from lipdverse based on query parameters
}
