% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryLipdverse.R
\name{queryLipdverse}
\alias{queryLipdverse}
\title{Query the LiPDverse}
\usage{
queryLipdverse(
  variable.name = NULL,
  archive.type = NULL,
  coord = c(-90, 90, -180, 180),
  age.min = NULL,
  age.max = NULL,
  pub.info = NULL,
  country = NULL,
  continent = NULL,
  ocean = FALSE,
  seasonality = NULL,
  season.not = NULL,
  interp.vars = NULL,
  interp.details = NULL,
  compilation = NULL,
  verbose = FALSE,
  skip.update = FALSE
)
}
\arguments{
\item{variable.name}{vector of variable names from c(d18O, d13C, treeRing, etc.)}

\item{archive.type}{vector of archive types. see unique(queryTable1$archiveType)}

\item{coord}{lat/lon extent of interest c(latMin, latMax, lonMin, lonMax)}

\item{age.min}{Sampling covers at least this recent extent, age BP}

\item{age.max}{Sampling covers at least this old extent, age BP}

\item{pub.info}{last name of author, DOI, unique word from title, etc. all lowercase, based on associated publications}

\item{country}{Coutnry origin of dataset from unique(queryTable1$country2), based on lat/lon}

\item{continent}{Continent origin of dataset from unique(queryTable1$continent), based on lat/lon}

\item{ocean}{Gather datasets from the marine environment, based on lat/lon}

\item{seasonality}{list of seasons where items within a list are treated with "AND" logic and separate lists are treated with "OR" logic ie. list(list("July", "August"), list("7,8"), list("summer"))}

\item{season.not}{seasons not desired with input format identical to seasonality}

\item{interp.vars}{vector of interpretation variables ie. c("SST", "upwelling"), see possible: unique(queryTable$interp_Vars)}

\item{interp.details}{vector of interpretation variables ie. c("sea@surface", "elNino"), see possible: unique(queryTable$interp_Details),}

\item{compilation}{#ompilation name as character vector from unique(queryTable$paleoData_mostRecentCompilations), eg. c("temp12k", "wnam")}

\item{verbose}{offer details of filters and get feedback from user}

\item{skip.update}{remove update prompt}
}
\value{
filtered query table
}
\description{
Query the LiPDverse
}
