% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{readLipd}
\alias{readLipd}
\title{Read LiPD files into R workspace}
\usage{
readLipd(
  path = NULL,
  jsonOnly = FALSE,
  parallel = FALSE,
  dont.load.ensemble = FALSE
)
}
\arguments{
\item{path}{A string specifying a file, url, or directory, or a vector of strings specifying files or urls. Alternatively, no entry (default) will open a selection windw..}

\item{jsonOnly}{Load data from json only (not lpd file? Typically only used for web connections)}

\item{parallel}{load data in parallel? Default = FALSE. Can greatly increase load time for large collections of data. Uses furrr, so you'll need to run future::plan() before calling the function. e.g. `future::plan(multisession,workers = 8)`}

\item{dont.load.ensemble}{This option doesn't load in ensemble data, but stores them in a temporary directory. If when that object is then written back out using `writeLipd()`, if that temporary directory still exists it will add the ensemble data back in. Default = FALSE}
}
\value{
D : LiPD dataset(s)
}
\description{
Read LiPD files into R workspace
}
\examples{
\dontrun{
#' read in multiple datasets - no path argument
D <- readLipd()   # choose option 'd' for directory

read in multiple datasets - with path argument
D <- readLipd("/Users/bobsmith/Desktop/lipd_files")

read in one dataset - no path argument
L <- readLipd()  # choose option "s" for single file

read in one dataset - with path argument
L <- readLipd("/Users/bobsmith/Desktop/lipd_files/dataset.lpd")
}
}
\author{
Chris Heiser

Nick McKay
}
\keyword{high-level}
