% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipdManipulation.R
\name{removeEnsembles}
\alias{removeEnsembles}
\title{Remove ensembles from a LiPD object}
\usage{
removeEnsembles(L)
}
\arguments{
\item{L}{a LiPD objec}
}
\value{
a LiPD object without paleo or chron ensembles
}
\description{
Remove ensembles from a LiPD object
}
