% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_helpers.R
\name{removeLayers}
\alias{removeLayers}
\title{Remove CSV and metadata layer from our lipd library. Also, remove empties}
\usage{
removeLayers(D, lpds)
}
\arguments{
\item{D}{LiPD Library}

\item{lpds}{List of LiPD files in the library}
}
\value{
D Modified lipd library
}
\description{
Remove CSV and metadata layer from our lipd library. Also, remove empties
}
\keyword{internal}

