% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{replaceLocalZipMD5}
\alias{replaceLocalZipMD5}
\title{Replace the query zip file MD5 sums after replacing the query table}
\usage{
replaceLocalZipMD5()
}
\description{
Replace the query zip file MD5 sums after replacing the query table
}
