% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_collapse.R
\name{rm_existing_tables}
\alias{rm_existing_tables}
\title{Remove tables that correspond to 'whichtables' type. These tables will be collapsed next and need a clean slate.}
\usage{
rm_existing_tables(d, pc, whichtables)
}
\arguments{
\item{d}{Metadata}

\item{pc}{paleoData or chronData}

\item{whichtables}{all summ meas ens}
}
\value{
d: Metadata
}
\description{
Remove tables that correspond to 'whichtables' type. These tables will be collapsed next and need a clean slate.
}
