% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_extract.R
\name{standardize_time_units}
\alias{standardize_time_units}
\title{Standardized time-units string}
\usage{
standardize_time_units(exponent, direction, u)
}
\arguments{
\item{exponent}{base-10 exponent of the units}

\item{direction}{"prograde" or "retrograde"}

\item{u}{lowercased original units string (may be NA)}
}
\value{
a standardized units string
}
\description{
Standardized time-units string
}
\keyword{internal}
