% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyStandardization.R
\name{standardizeKey}
\alias{standardizeKey}
\title{standardize terms automatically based on known synonyms}
\usage{
standardizeKey(lipdTS, key)
}
\arguments{
\item{key}{}
}
\value{
updated TS
}
\description{
standardize terms automatically based on known synonyms
}
