% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCstandardization.R
\name{standardizeQCsheetValues}
\alias{standardizeQCsheetValues}
\title{standardizeQCsheetValues}
\usage{
standardizeQCsheetValues(qcSheet)
}
\arguments{
\item{qcID}{}
}
\value{
stnadardized qc sheet
}
\description{
standardizeQCsheetValues
}
