% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{standardizeValue}
\alias{standardizeValue}
\title{standardize terms automatically based on known synonyms}
\usage{
standardizeValue(lipdTS, key = NA)
}
\arguments{
\item{key}{}
}
\value{
updated TS
}
\description{
standardize terms automatically based on known synonyms
}
