% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{summary.lipd}
\alias{summary.lipd}
\title{Full summary detail of single lipd}
\usage{
summary.lipd(object, ...)
}
\arguments{
\item{object}{a lipd object}

\item{...}{additional arguments passed to generic summary function}
}
\description{
Full summary detail of single lipd
}
