% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{summary.lipd_ts_tibble}
\alias{summary.lipd_ts_tibble}
\title{full summary of lipd ts tibble}
\usage{
\method{summary}{lipd_ts_tibble}(object, ...)
}
\arguments{
\item{object}{a lipd_ts_tibble object}

\item{...}{other arguments passed to lipdTSSummary()}
}
\value{
a summary table if return.table = TRUE
}
\description{
full summary of lipd ts tibble
}
