% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{summary.multi_lipd}
\alias{summary.multi_lipd}
\title{full summary of multi lipd}
\usage{
\method{summary}{multi_lipd}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
full summary of multi lipd
}
