% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{swap.geo.coordinates}
\alias{swap.geo.coordinates}
\title{An old bug caused some geo coordinates to be reversed. This will switch them back to normal.}
\usage{
swap.geo.coordinates(d)
}
\arguments{
\item{d}{Metadata}
}
\value{
d Modified metadata
}
\description{
An old bug caused some geo coordinates to be reversed. This will switch them back to normal.
}
\keyword{internal}

