% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulateTs}
\alias{tabulateTs}
\title{Create a data table from a lipd-ts-tibble object}
\usage{
tabulateTs(ts, time.var = "age")
}
\arguments{
\item{ts}{Any lipd class}

\item{time.var}{What's the name of the time var (default = "age")}
}
\description{
Create a data table from a lipd-ts-tibble object
}
