% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy2.R
\name{tidyTs}
\alias{tidyTs}
\title{create tidy data.frame from TS}
\usage{
tidyTs(TS, age.var = NA)
}
\arguments{
\item{TS}{a LiPD Timeseries object}

\item{age.var}{variable name for the time dimension}
}
\value{
a tidy data.frame
}
\description{
takes a TS object and turns it into a long, tidy, data.frame. Useful for data manipulation and analysis in the tidyverse and plotting. The opposite operation as untidyTs()
}
\seealso{
Other LiPD manipulation: 
\code{\link{pullTsVariable}()},
\code{\link{pushTsVariable}()},
\code{\link{untidyTs}()}
}
\concept{LiPD manipulation}
