% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_extract.R
\name{time_metadata_from_units}
\alias{time_metadata_from_units}
\title{Derive time datum/direction/exponent/units from a column's units and name}
\usage{
time_metadata_from_units(variableName, units)
}
\arguments{
\item{variableName}{The variableName of the time column}

\item{units}{The units of the time column}
}
\value{
list(timeUnits, timeDatum, timeDirection, timeExponent)
}
\description{
Derive time datum/direction/exponent/units from a column's units and name
}
\keyword{internal}
