% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{update_queryTable}
\alias{update_queryTable}
\title{update the query table with most recent lipdverse version}
\usage{
update_queryTable()
}
\description{
update the query table with most recent lipdverse version
}
