% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{updateNotes}
\alias{updateNotes}
\title{Update notes and paleoData_notes}
\usage{
updateNotes(
  lipdTS,
  key = NA,
  metadataChangesDF = NA,
  standardizeSynonymDF = NA,
  deleteTheseTS = NA
)
}
\arguments{
\item{lipdTS}{a LiPD TS object}

\item{metadataChangesDF}{a data frame listing the changes to the metadata}

\item{standardizeSynonymDF}{standardizing synonyms dataframe}
}
\value{
lipdTS
}
\description{
Update notes and paleoData_notes
}
