% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{updateQueryTable}
\alias{updateQueryTable}
\title{update the query table with most recent lipdverse version}
\usage{
updateQueryTable()
}
\description{
update the query table with most recent lipdverse version
}
