% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validLipd}
\alias{validLipd}
\title{Is this a valid LiPD}
\usage{
validLipd(L, allow.ensemble = TRUE)
}
\arguments{
\item{L}{a lipd object}

\item{allow.ensemble}{Check for ensemble variables as well}
}
\description{
Is this a valid LiPD
}
