% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warn_ensembles_in_paleo}
\alias{warn_ensembles_in_paleo}
\title{Warn people about writing ageEnsembles that have been mapped into paleoData. This is a common procedure in GeoChronR,
and thus will come up, however it can greatly increase the size of the LiPD file, and is easily and quickly replicated
upon loading with geoChronR::mapAgeEnsembleToPaleoData()}
\usage{
warn_ensembles_in_paleo(L, ignore.warnings)
}
\arguments{
\item{L}{LiPD file}

\item{ignore.warnings}{ignore warnings?}
}
\value{
Metadata
}
\description{
Warn people about writing ageEnsembles that have been mapped into paleoData. This is a common procedure in GeoChronR,
and thus will come up, however it can greatly increase the size of the LiPD file, and is easily and quickly replicated
upon loading with geoChronR::mapAgeEnsembleToPaleoData()
}
\keyword{internal}
