% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{writeLipd}
\alias{writeLipd}
\title{Write LiPD data onto disk as LiPD files}
\usage{
writeLipd(
  D,
  path = NULL,
  ignore.warnings = FALSE,
  removeNamesFromLists = FALSE,
  jsonOnly = FALSE,
  delete.saved.ensembles = FALSE
)
}
\arguments{
\item{D}{LiPD datasets  list}

\item{path}{Destination path  char}

\item{jsonOnly}{Write data to jsonld only? The data will be included and the json file might be large (Typically only used for web connections)}
}
\value{
none
}
\description{
Write LiPD data onto disk as LiPD files
}
\examples{
\dontrun{
# write - without path argument
writeLipd(D)

# write - with path argument
writeLipd(D, "/Users/bobsmith/Desktop/lipd_files")
}
}
\author{
Chris Heiser
}
\keyword{internal}
