% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{writeVersionToRoot}
\alias{writeVersionToRoot}
\title{write dataset version to a variable in the LiPD file}
\usage{
writeVersionToRoot(L)
}
\arguments{
\item{L}{A LiPD object}
}
\value{
A LiPD object
}
\description{
write dataset version to a variable in the LiPD file
}
