% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZIPmd5Local}
\alias{ZIPmd5Local}
\title{MD5 sums for Query Table Zip File}
\format{
## `ZIPmd5Local`
A set of 32-character hexadecimal letters and numbers:
}
\source{
<https://raw.githubusercontent.com/DaveEdge1/lipdverseQuery/main/ZIPmd5.txt>
}
\usage{
ZIPmd5Local
}
\description{
MD5 sums for Query Table Zip File
}
\keyword{datasets}
