% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_merge_data.R
\name{addCsvToMetadata}
\alias{addCsvToMetadata}
\title{Merge Main. Call the individual steps of merge for each file.}
\usage{
addCsvToMetadata(D, lpds)
}
\arguments{
\item{D}{LiPD Library}

\item{lpds}{LiPD names}
}
\value{
D Merged LiPD Library
}
\description{
Merge Main. Call the individual steps of merge for each file.
}
\keyword{internal}

