% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{createChangelog}
\alias{createChangelog}
\title{Create a changelog by comparing two LiPD files}
\usage{
createChangelog(
  Lold,
  Lnew,
  good.base = c("archiveType", "createdBy", "dataSetName", "dataSource", "notes",
    "originalDataUrl"),
  exclude.paleo = c("paleoData_meanValue12k", "paleoData_mostRecentCompilations",
    "paleoData_medianRes12k", "paleoData_tableName", "paleoData_values"),
  exclude.chron = c("chronData_values")
)
}
\arguments{
\item{exclude.chron}{}
}
\description{
Create a changelog by comparing two LiPD files
}
