% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{createSingleMarkdownChangelog}
\alias{createSingleMarkdownChangelog}
\title{Create markdown for a single entry}
\usage{
createSingleMarkdownChangelog(scl)
}
\arguments{
\item{scl}{}
}
\value{

}
\description{
Create markdown for a single entry
}
