% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{formatCitation}
\alias{formatCitation}
\title{Format citation from 'pub' portion of LiPD file}
\usage{
formatCitation(pub)
}
\description{
Format citation from 'pub' portion of LiPD file
}
\seealso{
Other summary: 
\code{\link{lipdSummary}()},
\code{\link{lipdTSSummary}()},
\code{\link{matchCols}()},
\code{\link{multiLipdSummary}()},
\code{\link{printCitations}()},
\code{\link{printSummaryData}()}
}
\author{
David Edge
}
\concept{summary}
