% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{getChangelog}
\alias{getChangelog}
\title{Get one instance from a LiPD changelog}
\usage{
getChangelog(L, version = "newest")
}
\arguments{
\item{version}{}
}
\value{

}
\description{
Get one instance from a LiPD changelog
}
