% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma2lipd.R
\name{getChronDataNeotoma2}
\alias{getChronDataNeotoma2}
\title{get chronData info from neotoma}
\usage{
getChronDataNeotoma2(site)
}
\arguments{
\item{site}{a single 'site' (not multiple 'sites') object generated by the 'neotoma2' package}
}
\value{
LiPD chronData section
}
\description{
get chronData info from neotoma
}
