% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma2lipd.R
\name{getPubNeotoma2}
\alias{getPubNeotoma2}
\title{get pub info from neotoma}
\usage{
getPubNeotoma2(site)
}
\arguments{
\item{site}{a single 'site' (not multiple 'sites') object generated by the 'neotoma2' package}
}
\value{
LiPD pub section
}
\description{
get pub info from neotoma
}
