% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_versions.R
\name{getVersion}
\alias{getVersion}
\title{Get the version number from metadata}
\usage{
getVersion(d)
}
\arguments{
\item{d}{LiPD Metadata}
}
\value{
version LiPD version number
}
\description{
Get the version number from metadata
}
\keyword{internal}

