% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lipds_indexing.R
\name{indexByName}
\alias{indexByName}
\title{Change index-by-number to index-by-variableName}
\usage{
indexByName(D, lpds)
}
\arguments{
\item{D}{The lipd library}

\item{lpds}{list of all LiPD files (no extension)}
}
\value{
D modified lipd library
}
\description{
Change index-by-number to index-by-variableName
}
\keyword{internal}

