% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipd2neotoma.R, R/neotoma2lipd.R
\name{lipd2neotoma}
\alias{lipd2neotoma}
\title{Convert lipd to neotoma}
\usage{
lipd2neotoma(L)

lipd2neotoma(L)
}
\arguments{
\item{L}{}
}
\value{
neotoma site

neotoma site
}
\description{
Convert lipd to neotoma

convert lipd format to neotoma
}
