% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{lipdTSSummary}
\alias{lipdTSSummary}
\title{Print a summary of the contents of a LiPD TS}
\usage{
lipdTSSummary(
  ts.object,
  print.length = 10,
  return.table = FALSE,
  add.variable = NULL
)
}
\arguments{
\item{ts.object}{a lipd_ts or lipd_ts_tibble}

\item{print.length}{length of printed summary table}

\item{return.table}{save the summary table}

\item{add.variable}{include additional variables from ts object in summary}
}
\value{
a summary table if return.table = TRUE
}
\description{
Print a summary of the contents of a LiPD TS
}
\seealso{
Other summary: 
\code{\link{formatCitation}()},
\code{\link{lipdSummary}()},
\code{\link{matchCols}()},
\code{\link{multiLipdSummary}()},
\code{\link{printCitations}()},
\code{\link{printSummaryData}()}
}
\author{
David Edge
}
\concept{summary}
