% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{multiLipdSummary}
\alias{multiLipdSummary}
\title{Print a summary of the contents of a Multi-LiPD}
\usage{
multiLipdSummary(
  multi.lipd,
  print.length = 20,
  time.units = "AD",
  return.table = FALSE
)
}
\arguments{
\item{L}{}
}
\description{
Print a summary of the contents of a Multi-LiPD
}
\seealso{
Other summary: 
\code{\link{formatCitation}()},
\code{\link{lipdSummary}()},
\code{\link{lipdTSSummary}()},
\code{\link{matchCols}()},
\code{\link{printCitations}()},
\code{\link{printSummaryData}()}
}
\author{
David Edge
}
\concept{summary}
