% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neotoma2lipd.R
\name{neotoma2lipd}
\alias{neotoma2lipd}
\title{Convert a neotoma dataset into a LiPD object}
\usage{
neotoma2lipd(site)
}
\arguments{
\item{site}{a single 'site' (not multiple 'sites') object generated by the 'neotoma2' package}
}
\value{
a LiPD object
}
\description{
Convert a neotoma dataset into a LiPD object
}
\examples{

site <- neotoma2::get_downloads(24)
L <- neotoma2lipd(site)

}
