% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{readLipd}
\alias{readLipd}
\title{Read LiPD files into R workspace}
\usage{
readLipd(path = NULL, jsonOnly = FALSE)
}
\arguments{
\item{path}{A string specifying a file, url, or directory, or a vector of strings specifying files or urls. Alternatively, no entry (default) will open a selection windw..}

\item{jsonOnly}{Load data from json only (not lpd file? Typically only used for web connections)}
}
\value{
D : LiPD dataset(s)
}
\description{
Read LiPD files into R workspace
}
\examples{
\dontrun{
#' read in multiple datasets - no path argument
D <- readLipd()   # choose option 'd' for directory

read in multiple datasets - with path argument
D <- readLipd("/Users/bobsmith/Desktop/lipd_files")

read in one dataset - no path argument
L <- readLipd()  # choose option "s" for single file

read in one dataset - with path argument
L <- readLipd("/Users/bobsmith/Desktop/lipd_files/dataset.lpd")
}
}
\author{
Chris Heiser

Nick McKay
}
\keyword{high-level}
