% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{summary.lipd_ts_tibble}
\alias{summary.lipd_ts_tibble}
\title{full simary of lipd ts tibble}
\usage{
\method{summary}{lipd_ts_tibble}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
full simary of lipd ts tibble
}
