% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{summary.multi_lipd}
\alias{summary.multi_lipd}
\title{full summary of multi lipd}
\usage{
\method{summary}{multi_lipd}(object, ...)
}
\arguments{
\item{object}{a multi_lipd object}

\item{...}{other arguments passed to multiLipdSummary()}
}
\value{
a summary table if return.table = TRUE
}
\description{
full summary of multi lipd
}
