% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{updateMetaDataFromStandardTables}
\alias{updateMetaDataFromStandardTables}
\title{grab metaData for given key from standard tables}
\usage{
updateMetaDataFromStandardTables(lipdTS, key)
}
\arguments{
\item{lipdTS}{a lipd TS object}

\item{key}{the key to check}
}
\value{
list
}
\description{
grab metaData for given key from standard tables
}
