% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{updateNotes}
\alias{updateNotes}
\title{Update notes and paleoData_notes}
\usage{
updateNotes(
  lipdTS,
  key = NA,
  metadataChangesDF = NA,
  standardizeSynonymDF = NA,
  deleteTheseTS = NA
)
}
\arguments{
\item{standardizeSynonymDF}{}
}
\value{
lipdTS
}
\description{
Update notes and paleoData_notes
}
